switched back to 32bit hashes, for the sake of performance experiment

+++ src/G2DHashMapTrie.cpp
@@ -249,7 +249,7 @@ namespace uva {
                     //1.1.2. Compute the m-gram hash
                     const uint8_t token_begin_idx = (ATrie<N>::m_query_ptr->level - level);
                     const uint8_t token_end_idx = (ATrie<N>::m_query_ptr->level - 1);
-                    uint64_t gram_hash = ATrie<N>::m_query_ptr->suffix_hash(token_begin_idx);
+                    uint32_t gram_hash = ATrie<N>::m_query_ptr->suffix_hash(token_begin_idx);
                     LOG_DEBUG << "The " << level << "-gram: " << tokensToString(ATrie<N>::m_query_ptr->tokens,
                             token_begin_idx, token_end_idx) << " hash is " << gram_hash << END_LOG;
 
@@ -312,7 +312,7 @@ namespace uva {
                     //1.1.2. Compute the hash value for the back off M-gram
                     const uint8_t token_begin_idx = (ATrie<N>::m_query_ptr->level - 1) - level;
                     const uint8_t token_end_idx = (ATrie<N>::m_query_ptr->level - 2);
-                    uint64_t gram_hash = ATrie<N>::m_query_ptr->sub_hash(
+                    uint32_t gram_hash = ATrie<N>::m_query_ptr->sub_hash(
                             token_begin_idx, token_end_idx);
                     LOG_DEBUG << "The: " << tokensToString(ATrie<N>::m_query_ptr->tokens, token_begin_idx, token_end_idx)
                             << " " << level << "-gram back-off hash is " << gram_hash << END_LOG;
